omd: move sitename_must_be_valid
`omd/packages/omd/omdlib/main.py` is way too big.

Also replace `bail_out` by `sys.exit`.
